Add CODEOWNERS & split commands by Subsystems
