+++ src/org/usfirst/frc/team334/robot/commands/Drivetrain/FillAirTanksCommand.java
@@ -1,13 +1,13 @@
-package org.usfirst.frc.team334.robot.commands;
+package org.usfirst.frc.team334.robot.commands.Drivetrain;
 
-import org.usfirst.frc.team334.robot.Robot;
 
 import edu.wpi.first.wpilibj.command.Command;
 
-public class SetElevatorToScaleCommand extends Command {
+public class FillAirTanksCommand extends Command {
 	
-	public SetElevatorToScaleCommand() {
-		requires(Robot.sElevator);
+	public FillAirTanksCommand() {
+		// Use requires() here to declare subsystem dependencies
+		// requires(Robot.sPneumatics);
 	}
 
 	// Called just before this Command runs the first time
@@ -19,7 +19,7 @@ public class SetElevatorToScaleCommand extends Command {
 	// Called repeatedly when this Command is scheduled to run
 	@Override
 	protected void execute() {
-		
+		System.out.println("FILLING AIRTANKS");
 	}
 
 	// Make this return true when this Command no longer needs to run execute()
