+++ src/org/usfirst/frc/team334/robot/commands/Elevator/SetElevatorToIntakeCommand.java
@@ -1,19 +1,19 @@
-package org.usfirst.frc.team334.robot.commands;
+package org.usfirst.frc.team334.robot.commands.Elevator;
 
 import org.usfirst.frc.team334.robot.Robot;
 
 import edu.wpi.first.wpilibj.command.Command;
 
-public class ShiftToLowSpeedTransmissionCommand extends Command {
+public class SetElevatorToIntakeCommand extends Command {
 	
-	public ShiftToLowSpeedTransmissionCommand() {
-		 requires(Robot.sPneumatics);
+	public SetElevatorToIntakeCommand() {
+		requires(Robot.sElevator);
 	}
 
 	// Called just before this Command runs the first time
 	@Override
 	protected void initialize() {
-		Robot.sPneumatics.setLowSpeedTransmission();
+		
 	}
 
 	// Called repeatedly when this Command is scheduled to run
@@ -25,7 +25,7 @@ public class ShiftToLowSpeedTransmissionCommand extends Command {
 	// Make this return true when this Command no longer needs to run execute()
 	@Override
 	protected boolean isFinished() {
-		return true;
+		return false;
 	}
 
 	// Called once after isFinished returns true
