+++ src/org/usfirst/frc/team334/robot/commands/Intake/ReleasePowerCubeCommand.java
@@ -1,4 +1,4 @@
-package org.usfirst.frc.team334.robot.commands;
+package org.usfirst.frc.team334.robot.commands.Intake;
 
 import org.usfirst.frc.team334.robot.Robot;
